+++ nbproject/private/private.properties
@@ -1,2 +1,6 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\sayan\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+do.depend=false
+do.jar=true
+javac.debug=true
+javadoc.preview=true
+user.properties.file=C:\\Users\\prate\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
