+++ nbproject/project.properties
@@ -1,9 +1,10 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
+application.title=Web crawler
+application.vendor=prate
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -26,18 +27,12 @@ dist.archive.excludes=
 dist.dir=dist
 dist.jar=${dist.dir}/Web_crawler.jar
 dist.javadoc.dir=${dist.dir}/javadoc
+endorsed.classpath=
 excludes=
-file.reference.crawler-jsoup-1.10.2=src\\web\\crawler\\jsoup-1.10.2
-file.reference.jsoup-1.10.2-javadoc.jar=src\\web\\crawler\\jsoup-1.10.2-javadoc.jar
-file.reference.jsoup-1.10.2-sources.jar=src\\web\\crawler\\jsoup-1.10.2-sources.jar
-file.reference.jsoup-1.10.2.jar=src\\web\\crawler\\jsoup-1.10.2.jar
 includes=**
 jar.compress=false
 javac.classpath=\
-    ${file.reference.crawler-jsoup-1.10.2}:\
-    ${file.reference.jsoup-1.10.2-javadoc.jar}:\
-    ${file.reference.jsoup-1.10.2-sources.jar}:\
-    ${file.reference.jsoup-1.10.2.jar}
+    ${libs.MyJsoup.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
