+++ src/web/crawler/SpiderLeg.java
@@ -1,58 +1,23 @@
-import java.util.Scanner;
-import java.util.Set;
-import java.io.IOException;
-import java.util.List;
-import java.util.HashSet;
-import java.util.LinkedList;
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package web.crawler;
 
+import java.io.IOException;
+import java.util.LinkedList;
+import java.util.List;
 import org.jsoup.Connection;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
-public class Spider
-{
-    private static final int MAX_PAGES_TO_SEARCH=15;
-    private Set<String> pagesVisited=new HashSet<String>();
-    private List<String> pagesToVisit=new LinkedList<String>();
-    
-    private String nextUrl()
-    {
-        String nextUrl;
-        do
-        {
-            nextUrl=this.pagesToVisit.remove(0);
-        }while(this.pagesVisited.contains(nextUrl));
-        return nextUrl;
-    }
-    public void search(String url, String searchWord)
-    {
-        while(this.pagesVisited.size()<MAX_PAGES_TO_SEARCH)
-        {
-            String currentUrl;
-            SpiderLeg leg=new SpiderLeg();
-            if(this.pagesToVisit.isEmpty())
-            {
-                currentUrl=url;
-                this.pagesVisited.add(url);
-            }
-            else
-            {
-                currentUrl=this.nextUrl();
-            }
-            leg.crawl(currentUrl);
-            boolean success=leg.searchForWord(searchWord);
-            if(success)
-            {
-                System.out.println(String.format("Success! word %s found at url: %s ",searchWord, currentUrl));
-                break;
-            }
-            this.pagesToVisit.addAll(leg.getLinks());
-        }
-        System.out.println(String.format("Task completed! %s pages visited",this.pagesVisited.size()));
-    }
-}
+/**
+ *
+ * @author prate
+ */
 public class SpiderLeg
 {
     private static final String USER_AGENT =
@@ -86,6 +51,7 @@ public class SpiderLeg
         {
             System.out.println("Error in out HTTP request " + ioe);
         }
+        return true;
     }
     public boolean searchForWord(String searchWord)
     {
@@ -104,16 +70,3 @@ public class SpiderLeg
          return this.links;
      }
   }
-
-public class SpiderTest
-{
-    public static void main(String[] args)
-    {
-        Spider spider=new Spider();
-        System.out.println("Enter start url, Search word");
-        Scanner s=new Scanner(System.in);
-        String url=s.nextLine();
-        String word=s.nextLine();
-        spider.search(url, word);
-    }
-}
